# laravel/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REST API студии ремонта (список методов, для работы с таблицами ремонтов).",
        "description": "Документация представляет собой REST API методы для взаимойствия с базой данных ",
        "version": "1.0.0"
    },
    "paths": {
        "/api/elements": {
            "get": {
                "summary": "Список всех услуг, входящих в комплексные ремонты",
                "description": "Возвращает список всех услуг, входящих в комплексные ремонты.",
                "operationId": "2b0a99d189b3930e814455926c208ce7",
                "responses": {
                    "200": {
                        "description": "Услуги"
                    },
                    "500": {
                        "description": "Записи не найдены"
                    }
                }
            }
        },
        "/api/catalog": {
            "get": {
                "summary": "Список комплексных ремонтов",
                "description": "Возвращает весь список комплексных ремонтов.",
                "operationId": "d263c9fec0b711a158012190796ff8d0",
                "responses": {
                    "200": {
                        "description": "Комплексные ремонты"
                    },
                    "500": {
                        "description": "Записи не найдены"
                    }
                }
            }
        },
        "/api/search?q={string}": {
            "get": {
                "summary": "Поиск элементов",
                "description": "Возвращает весь список комплексных ремонтов, который соответствует подстроке, введенной в запросе.",
                "operationId": "f05ad2c123892bf2907fdc9102349f28",
                "parameters": [
                    {
                        "name": "string",
                        "in": "path",
                        "description": "После GET-параметра q необходимо ввести название интересующего комплексного ремонта.",
                        "required": true,
                        "example": "кухни"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Комплексные ремонты"
                    },
                    "400": {
                        "description": "Строка оказалась пустой, или не был обнаружен GET-параметр q."
                    },
                    "404": {
                        "description": "Не удалось найти записи по данной строке"
                    }
                }
            }
        },
        "/api/catalog/{id}": {
            "get": {
                "summary": "Список элементов по комплексному ремонту",
                "description": "Возвращает все услуги, которые входят в состав комплексного ремонта. ",
                "operationId": "c4722204b713cd6611d7e7fd4b78f594",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id записи с комплексным ремонтом",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список услуг, входящих в состав комплексного"
                    },
                    "404": {
                        "description": "Nothing was found"
                    }
                }
            }
        }
    }
}